- package lock version 2 - added whitespace between some template blocks - added some comment to describe what some components do - removed some accessors - renamed search.vue -> new-search.vue - updated some component IDs - fixed template errors in pick-request-type
