+++ src/components/new-request/main-container.vue
@@ -2,6 +2,7 @@
   <v-container fluid class="white rounded copy-normal pa-10">
     <v-row justify="space-between" class="submission-tabs-top" no-gutters>
       <slot name="container-header" />
+
       <v-col cols="auto" class="py-0" v-if="displayedComponent !== 'Success'">
         <button id="back-to-search-btn"
                 class="modal-activator link-std-sans-ul"
@@ -17,7 +18,8 @@
         </button>
       </v-col>
     </v-row>
-    <slot name="content"/>
+
+    <slot name="content" />
   </v-container>
 </template>
