+++ tests/unit/new-search.spec.ts
@@ -1,5 +1,5 @@
 import { createLocalVue, mount, shallowMount } from '@vue/test-utils'
-import NewSearch from '@/components/new-request/search.vue'
+import NewSearch from '@/components/new-request/new-search.vue'
 // import newReqModule from '@/store/new-request-module'
 import Vuetify from 'vuetify'
